requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
Django==4.0.3
django-filter==21.1
django-silk==4.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gprof2dot==2021.2.21
idna==3.3
Jinja2==3.0.3
jwt==1.3.1
MarkupSafe==2.1.0
Pillow==9.0.1
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
